RR-526 - Refactor CIAG Induction mappers into 1 source file
